Add detail to Octokit API Exceptions (#184)
* Update resource string

* Add more detailed error message to Octokit validation errors

* Nit string swaps
