+++ GitHubExtension/DeveloperId/DeveloperIdProvider.cs
@@ -206,10 +206,6 @@ public class DeveloperIdProvider : IDeveloperIdProvider
 
     private void RestoreDeveloperIds(IEnumerable<string> loginIdsAndUrls)
     {
-        // We take loginIds or Urls here because in older versions of DevHome, we used loginIds to save credentials.
-        // In newer versions, we use Urls to save credentials.
-        // So, we need to check if loginId is currently used to save credential, and if so, replace it with URL.
-        // This is a temporary fix, and we should replace this logic once we are sure that most users have updated to newer versions of DevHome.
         foreach (var loginIdOrUrl in loginIdsAndUrls)
         {
             // Since GitHub loginIds cannot contain /, and URLs would, this is sufficient to differentiate between
